Move constants to a constants module for now and fix redis url
