+++ wazimap_ng/constants.py
@@ -1,0 +1,36 @@
+DENOMINATOR_CHOICES = (
+    ('absolute_value', 'Absolute value'),
+    ('subindicators', 'Sub-indicators'),
+    ('sibling', 'Sibling'),
+)
+
+PERMISSION_TYPES = (
+    ('private', 'Private'),
+    ('public', 'Public'),
+)
+
+
+QUANTITATIVE = "quantitative"
+QUALITATIVE = "qualitative"
+
+DATASET_CONTENT_TYPES = (
+    ('quantitative', 'Quantitative'),
+    ('qualitative', 'Qualitative')
+)
+
+PI_CONTENT_TYPE = (
+    ('indicator', 'Indicator'),
+    ('html', 'HTML')
+)
+
+PI_CHART_TYPE = (
+    ('bar', 'Bar Chart'),
+    ('line', 'Line Chart')
+)
+
+PI_CHOROPLETH_RANGE_TYPE = (
+    ('by_subindicator', 'By Subindicator'),
+    ('by_indicator', 'By Indicator')
+)
+
+STAFF_GROUPS = ["ProfileAdmin", "DataAdmin"]
