+++ wazimap_ng/datasets/hooks.py
@@ -10,7 +10,7 @@ from .models import Indicator
 from django.core import mail
 from django.template.loader import render_to_string
 from django.utils.html import strip_tags
-from wazimap_ng.config.common import STAFF_EMAIL_ADDRESS
+from django.conf import settings
 
 import json
 
@@ -159,7 +159,7 @@ def process_task_info(task):
                 html_message = render_to_string('emailTemplates/indicator_extraction_task_notification.html', context)
 
             plain_message = strip_tags(html_message)
-            from_email = F'From <{STAFF_EMAIL_ADDRESS}>'
+            from_email = F'From <{settings.STAFF_EMAIL_ADDRESS}>'
             to = user.email
             mail.send_mail(subject, plain_message, from_email, [to], html_message=html_message)
 
@@ -187,7 +187,7 @@ def custom_admin_notification(session, notification_type, message, task_id=None)
     notifications are stored in session and show to user when user refreshes page.
 
     A valid session object must be passed to this function with notification type and message
-    
+
     Type of notifications:
         * success
         * info
